+++ src/server.js
@@ -1,7 +1,6 @@
 const { EventEmitter } = require('events')
 const { createDeserializer, createSerializer } = require('./transforms/serializer')
 const { Player } = require('./serverPlayer')
-const { RakServer } = require('./rak')
 const { sleep } = require('./datatypes/util')
 const { ServerAdvertisement } = require('./server/advertisement')
 const Options = require('./options')
@@ -12,6 +11,9 @@ class Server extends EventEmitter {
     super()
     this.options = { ...Options.defaultOptions, ...options }
     this.validateOptions()
+
+    this.RakServer = require('./rak')(this.options.useNativeRaknet).RakServer
+
     this.serializer = createSerializer(this.options.version)
     this.deserializer = createDeserializer(this.options.version)
     this.advertisement = new ServerAdvertisement(this.options.motd)
@@ -66,7 +68,7 @@ class Server extends EventEmitter {
   }
 
   async listen (host = this.options.host, port = this.options.port) {
-    this.raknet = new RakServer({ host, port }, this)
+    this.raknet = new this.RakServer({ host, port }, this)
     try {
       await this.raknet.listen()
     } catch (e) {