+++ test/vanilla.test.js
@@ -4,7 +4,8 @@ const { clientTest } = require('./vanilla')
 const { Versions } = require('../src/options')
 
 describe('vanilla server test', function () {
-  this.timeout(220 * 1000)
+  const vcount = Object.keys(Versions).length
+  this.timeout(vcount * 80 * 1000)
 
   for (const version in Versions) {
     it('client spawns ' + version, async () => {